adding Static folder with 4 components and moving code there from ZipMarkdownToHtml
