+++ src/components/ZipMarkdownToHtml/Static/LocalConversionDescription.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const LocalConversionDescription = () => {
+  return (
+    <section>
+      <h2>Local Conversion</h2>
+      <p>The same can be done locally from your computer. UPLOAD - CONVERT - DOWNLOAD</p>
+    </section>
+  );
+};
+
+export default LocalConversionDescription;
